requirements.txt
google-cloud-secret-manager
google-crc32c
discord.py
openai